Add support for response builders on IRequestConfiguration
This allows consumers to provide builders on a per-request basis,
required in ingest for streaming bulk response.

Fixes potential null ref exceptions when disposing streams.

This also pins the SDK to 8.x until we're ready for 9.x
