+++ src/Elastic.Transport/Components/Pipeline/RequestData.cs
@@ -7,7 +7,6 @@ using System.Collections.Generic;
 using System.Collections.Specialized;
 using System.Security.Cryptography.X509Certificates;
 using Elastic.Transport.Extensions;
-using Elastic.Transport.Products;
 
 namespace Elastic.Transport;
 
@@ -41,7 +40,7 @@ public sealed record RequestData
 		ProxyPassword = global.ProxyPassword;
 		DisableAutomaticProxyDetection = global.DisableAutomaticProxyDetection;
 		UserAgent = global.UserAgent;
-		ResponseBuilders = global.ResponseBuilders;
+		ResponseBuilders = local?.ResponseBuilders ?? global.ResponseBuilders;
 		ProductResponseBuilders = global.ProductRegistration.ResponseBuilders;
 
 		KeepAliveInterval = (int)(global.KeepAliveInterval?.TotalMilliseconds ?? 2000);
@@ -92,11 +91,6 @@ public sealed record RequestData
 		}
 	}
 
-	/// <inheritdoc cref="ITransportConfiguration.ResponseBuilders"/>
-	public IReadOnlyCollection<IResponseBuilder> ProductResponseBuilders { get; }
-
-	/// <inheritdoc cref="ITransportConfiguration.ResponseBuilders"/>
-	public IReadOnlyCollection<IResponseBuilder> ResponseBuilders { get; }
 
 	/// <inheritdoc cref="ITransportConfiguration.MemoryStreamFactory"/>
 	public MemoryStreamFactory MemoryStreamFactory { get; }
@@ -168,4 +162,8 @@ public sealed record RequestData
 	public bool DisableSniff { get; }
 	/// <inheritdoc cref="IRequestConfiguration.DisablePings"/>
 	public bool DisablePings { get; }
+	/// <inheritdoc cref="IRequestConfiguration.ResponseBuilders"/>
+	public IReadOnlyCollection<IResponseBuilder> ProductResponseBuilders { get; }
+	/// <inheritdoc cref="IRequestConfiguration.ResponseBuilders"/>
+	public IReadOnlyCollection<IResponseBuilder> ResponseBuilders { get; }
 }
