+++ src/Elastic.Transport/Components/TransportClient/HttpWebRequestInvoker.cs
@@ -190,8 +190,8 @@ public class HttpWebRequestInvoker : IRequestInvoker
 			}
 			else
 			{
-				responseStream.Dispose();
-				receivedResponse.Dispose();
+				responseStream?.Dispose();
+				receivedResponse?.Dispose();
 			}
 
 			if (OpenTelemetry.CurrentSpanIsElasticTransportOwnedAndHasListeners && (Activity.Current?.IsAllDataRequested ?? false))
@@ -208,8 +208,8 @@ public class HttpWebRequestInvoker : IRequestInvoker
 		catch
 		{
 			// if there's an exception, ensure we always release the stream and response so that the connection is freed.
-			responseStream.Dispose();
-			receivedResponse.Dispose();
+			responseStream?.Dispose();
+			receivedResponse?.Dispose();
 			throw;
 		}
 	}
